Added plugin system on planner
