+++ ltl_automaton_msgs/cfg/ltl_automaton_dynparam.cfg
@@ -5,6 +5,6 @@ from dynamic_reconfigure.parameter_generator_catkin import *
 
 gen = ParameterGenerator()
 
-gen.add("re_plan_hil_param",    bool_t,    0, "Toggles ltl re-plannig to avoid re-planning during hil control", True)
+gen.add("replan_on_unplanned_move",    bool_t,    0, "Toggles replanning when received TS state in not the expected next one in the plan", True)
 
 exit(gen.generate(PACKAGE, "ltl_automaton_planner", "LTLAutomatonDP"))
