+++ ltl_automaton_planner/launch/ltl_planner_testdynparam.launch
@@ -4,7 +4,7 @@
   <arg name="initial_ts_state_from_agent" default="False"/>
 
   <!-- Run planner node -->
-  <node name="ltl_planner" pkg="ltl_automaton_planner" type="planner.py" output="screen">
+  <node name="ltl_planner" pkg="ltl_automaton_planner" type="planner_node.py" output="screen">
     <param name="initial_ts_state_from_agent" type="bool" value="$(arg initial_ts_state_from_agent)" />
   </node>
