+++ ltl_automaton_planner/src/ltl_automaton_planner/boolean_formulas/parser.py
