+++ ltl_automaton_planner/src/ltl_automaton_planner/ltl_tools/discrete_plan.py
@@ -2,7 +2,7 @@
 
 from product import ProdAut_Run
 from collections import defaultdict
-from networkx import dijkstra_predecessor_and_distance, has_path
+from networkx import dijkstra_predecessor_and_distance
 import time 
 
 
@@ -271,21 +271,6 @@ def dijkstra_revise_once(product, run_segment, broken_edge_index):
     for (bridge, cost) in dijkstra_targets(product, run_segment[broken_edge_index-1], set([run_segment[-1]])):
         new_run_segment = run_segment[0:(broken_edge_index-1)] + bridge
         return new_run_segment
-
-def has_path_to_accept(product, f_s):
-        for acc in product.graph['accept']:
-                if has_path(product, f_s, acc):
-                        return True
-        return False
-
-def has_path_to_accept_with_cycle(product, f_s):
-        print "---- Accepting states with cycles are ------"
-        print product.graph['accept_with_cycle']
-        print "------------------------------"
-        for acc in product.graph['accept_with_cycle']:
-            if has_path(product, f_s, acc):
-                return True
-            return False
 
 def compute_path_ac_d(product, path):
     ac_d = 0
