+++ ltl_automaton_planner/src/ltl_automaton_planner/ltl_tools/ltl2ba
